Rename handler to nodepool
