+++ service/controller/resource/nodepool/create.go
@@ -1,4 +1,4 @@
-package instance
+package nodepool
 
 import (
 	"context"
