+++ service/controller/resource/nodepool/create_cordon_old_workers.go
@@ -1,4 +1,4 @@
-package instance
+package nodepool
 
 import (
 	"context"
