+++ service/controller/resource/nodepool/create_deployment_uninitialized.go
@@ -1,4 +1,4 @@
-package instance
+package nodepool
 
 import (
 	"context"
