+++ service/controller/resource/nodepool/create_drain_old_worker_nodes.go
@@ -1,4 +1,4 @@
-package instance
+package nodepool
 
 import (
 	"context"
