+++ service/controller/resource/nodepool/create_scale_workers.go
@@ -1,4 +1,4 @@
-package instance
+package nodepool
 
 import (
 	"context"
@@ -10,7 +10,7 @@ import (
 	"github.com/giantswarm/azure-operator/v4/service/controller/internal/state"
 	"github.com/giantswarm/azure-operator/v4/service/controller/internal/vmsscheck"
 	"github.com/giantswarm/azure-operator/v4/service/controller/key"
-	"github.com/giantswarm/azure-operator/v4/service/controller/resource/instance/scalestrategy"
+	"github.com/giantswarm/azure-operator/v4/service/controller/resource/nodes/scalestrategy"
 )
 
 // The goal of scaleUpWorkerVMSSTransition is to double the desired number
