+++ service/controller/resource/nodepool/create_wait_for_nodes_to_become_ready.go
@@ -1,4 +1,4 @@
-package instance
+package nodepool
 
 import (
 	"context"
