+++ service/controller/resource/nodepool/status.go
@@ -1,4 +1,4 @@
-package instance
+package nodepool
 
 import (
 	"context"
