+++ service/controller/resource/nodepool/template/main.json
