+++ service/controller/resource/nodepool/template/template.go
