+++ service/controller/resource/nodes/scalestrategy/interface.go
