+++ service/controller/resource/nodes/scalestrategy/quick.go
